Fix signedness issue in loop and enhance compiler flags
Changed the loop index type from `int` to `size_t` to match `line.size()` and prevent potential signed/unsigned comparison warnings. Added comprehensive compiler warning flags to CMake configuration for improved code quality and stricter compilation checks.

+++ tyan.h
@@ -123,7 +123,7 @@ namespace tyan {
             std::string tmp;
             std::ostringstream oss;
             oss << std::string(ThreadContext::get().depth() << 1, ' ');
-            for (int i = 0; i < line.size(); i++) {
+            for (size_t i = 0; i < line.size(); i++) {
                 char c = line[i];
                 if (VAR_CHARSET.count(c)) {
                     tmp += c;
